Delete remnants of old submodule directories
