--- core/src/main/resources/log4j.properties
@@ -1,17 +1,0 @@
-# suppress inspection "UnusedProperty" for whole file
-
-# Root logger option
-log4j.rootLogger=WARN, CONSOLE
-log4j.logger.org=WARN
-log4j.logger.com=WARN
-log4j.logger.java=WARN
-log4j.logger.javax=WARN
-log4j.logger.scala=WARN
-
-# Direct log messages to stdout
-log4j.appender.CONSOLE=org.apache.log4j.ConsoleAppender
-log4j.appender.CONSOLE.layout=org.apache.log4j.PatternLayout
-log4j.appender.CONSOLE.layout.ConversionPattern=%d{yyyy-MM-dd'T'HH:mm:ss.SSSX} %5p %c{1}:%L - %m%n
-
-# Package specific logging configurations
-log4j.logger.org.springframework=WARN
