+++ src/test/load/scala/fi/vm/sade/hakurekisteri/integration/valintatulos/ValintaTulosLoadSpec.scala
