+++ src/test/load/scala/fi/vm/sade/hakurekisteri/kkhakija/KkhakijaResourceLoadSpec.scala
