+++ src/test/load/scala/fi/vm/sade/hakurekisteri/oppija/OppijaResourceLoadSpec.scala
