--- web/src/it/resources/log4j.properties
@@ -1,10 +1,0 @@
-# suppress inspection "UnusedProperty" for whole file
-# Root logger option
-log4j.rootLogger=INFO, NOLOG
-log4j.appender.NOLOG = org.apache.log4j.varia.NullAppender
-
-# Direct log messages to stdout
-log4j.appender.CONSOLE=org.apache.log4j.ConsoleAppender
-log4j.appender.CONSOLE.layout=org.apache.log4j.PatternLayout
-log4j.appender.CONSOLE.layout.ConversionPattern=%d{yyyy-MM-dd'T'HH:mm:ss.SSSX} %5p %c{1}:%L - %m%n
-
