--- web/src/test/resources/log4j.properties
@@ -1,19 +1,0 @@
-# suppress inspection "UnusedProperty" for whole file
-# Root logger option
-log4j.rootLogger=INFO, NOLOG
-log4j.appender.NOLOG = org.apache.log4j.varia.NullAppender
-
-# Direct log messages to stdout
-log4j.appender.CONSOLE=org.apache.log4j.ConsoleAppender
-log4j.appender.CONSOLE.layout=org.apache.log4j.PatternLayout
-log4j.appender.CONSOLE.layout.ConversionPattern=%d{yyyy-MM-dd'T'HH:mm:ss.SSSX} %5p %c{1}:%L - %m%n
-
-# Audit logger configuration
-log4j.logger.fi.vm.sade.auditlog.Audit=INFO, audit
-log4j.appender.audit=org.apache.log4j.RollingFileAppender
-log4j.appender.audit.File=./logs/auditlog_hakurekisteri_test.log
-log4j.appender.audit.MaxFileSize=20MB
-log4j.appender.audit.MaxBackupIndex=20
-log4j.appender.audit.layout=org.apache.log4j.PatternLayout
-log4j.appender.audit.layout.ConversionPattern=%m%n
-log4j.appender.audit.encoding=UTF-8
